Propose municipal internet fund
This fund would provide financing to local municipalities to build fiber networks, as opposed to depending on private companies to build these networks. Over time the investment pays for itself and allows for open access to digital infrastructure and more competition among internet providers, leading to more options and lower costs for consumers.